requirements.txt
streamlit==1.37.1
langchain==0.2.14
langchain-openai==0.1.7
python-dotenv==1.0.1
openai==1.40.0

